# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptCreateRequest": {
        "properties": {
          "problemVersionId": {
            "format": "uuid",
            "title": "Problemversionid",
            "type": "string"
          }
        },
        "required": [
          "problemVersionId"
        ],
        "title": "AttemptCreateRequest",
        "type": "object"
      },
      "AttemptResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "problemVersionId": {
            "format": "uuid",
            "title": "Problemversionid",
            "type": "string"
          },
          "status": {
            "enum": [
              "draft",
              "submitted"
            ],
            "title": "Status",
            "type": "string"
          },
          "studyProfileId": {
            "format": "uuid",
            "title": "Studyprofileid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "studyProfileId",
          "problemVersionId",
          "status",
          "createdAt"
        ],
        "title": "AttemptResponse",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "TranscriptConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "transcriptHash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Transcripthash",
            "type": "string"
          },
          "transcriptVersionId": {
            "format": "uuid",
            "title": "Transcriptversionid",
            "type": "string"
          }
        },
        "required": [
          "transcriptVersionId",
          "transcriptHash"
        ],
        "title": "TranscriptConfirmRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Math Coach API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/attempts": {
      "post": {
        "operationId": "post_attempt_api_v1_attempts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed or is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Owned resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request or upload validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An upstream structured response was invalid"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A required service is unavailable"
          }
        },
        "summary": "Post Attempt",
        "tags": [
          "attempts"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}": {
      "get": {
        "operationId": "get_attempt_api_v1_attempts__attempt_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed or is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Owned resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request or upload validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An upstream structured response was invalid"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A required service is unavailable"
          }
        },
        "summary": "Get Attempt",
        "tags": [
          "attempts"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/confirm-transcript": {
      "post": {
        "operationId": "post_transcript_confirmation_api_v1_attempts__attempt_id__confirm_transcript_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptConfirmationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed or is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Owned resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request or upload validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An upstream structured response was invalid"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A required service is unavailable"
          }
        },
        "summary": "Post Transcript Confirmation",
        "tags": [
          "transcription"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/hints/next": {
      "post": {
        "operationId": "post_next_hint_api_v1_attempts__attempt_id__hints_next_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NextHintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextHintResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed or is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Owned resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request or upload validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An upstream structured response was invalid"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A required service is unavailable"
          }
        },
        "summary": "Post Next Hint",
        "tags": [
          "static-journey"
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/mock-evaluation": {
      "post": {
        "operationId": "post_mock_evaluation_api_v1_attempts__attempt_id__mock_evaluation_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MockEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MockEvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed or is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Owned resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request or upload validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An upstream structured response was invalid"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A required service is unavailable"
          }
        },
        "summary": "Post Mock Evaluation",
        "tags": [
          "static-journey"
        ]
      }
    },
...
    "/api/v1/attempts/{attempt_id}/transcripts": {
      "post": {
        "operationId": "post_transcript_version_api_v1_attempts__attempt_id__transcripts_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptVersionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication failed or is required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Owned resource was not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request or upload validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unexpected server error occurred"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An upstream structured response was invalid"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "A required service is unavailable"
          }
        },
        "summary": "Post Transcript Version",
        "tags": [
          "transcription"
        ]
      }
    },
...
  }
}